bigint

+++ leaderboard_transfer/Transfertokens/koiosf_transfertokens.mjs
@@ -118,7 +118,7 @@ async function SendTransaction() {
         if (name == "Titan") {
             for (var i=0;i<sendlist.length;i++) {     
                 console.log(parseInt(sendlist[i][1]) * 10**18);           
-                var transaction = await contracttoken.methods.transfer(sendlist[i][0], (parseInt(sendlist[i][1]) * 10**18)).send({from: globalaccounts[0]});
+                var transaction = await contracttoken.methods.transfer(sendlist[i][0], BigInt(parseInt(sendlist[i][1]) * 10**18)).send({from: globalaccounts[0]});
                 console.log(transaction)
             }
         }
